pyproject.toml
[tool.poetry]
name = "teached"
version = "0.1.0"
description = "online learning platform"
authors = ["Mohamed Nesredin <m.n.kaizen@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "example.com"
documentation = ""
repository = "https://github.com/Mohamed-Kaizen/teached/"
keywords = ["teached"]

[tool.poetry.dependencies]
python = "^3.7"
confusable_homoglyphs = "^3.2.0"
fastapi = {extras = ["all"], version = "^0.55.1"}
loguru = "^0.5.0"
importlib_metadata = {version = "^1.6.0", python = "<3.8"}
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
pendulum = "^2.1.0"
pyjwt = "^1.7.1"
python-dotenv = "^0.13.0"
tortoise-orm = {extras = ["accel"], version = "^0.16.13"}
typer = {extras = ["all"], version = "^0.2.1"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
codecov = "^2.1.4"
coverage = {extras = ["toml"], version = "^5.1"}
darglint = "^1.4.0"
flake8 = "^3.8.2"
flake8-annotations = "^2.1.0"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.0"
flake8-bugbear = "^20.1.4"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
isort = {extras = ["pyproject"], version = "^4.3.21"}
mkdocs = "^1.1.2"
mkdocs-material = "^5.2.3"
mkdocstrings = "^0.11.3"
mkdocs-minify-plugin = "^0.3.0"
mkdocs-git-revision-date-localized-plugin = "^0.5.2"
mypy = "^0.770"
pytest = "^5.4.2"
pytest-cov = "^2.9.0"
pytype = {version = "^2020.5.13", python = "3.7"}
safety = "^1.9.0"
typeguard = "^2.9.1"
xdoctest = "^0.12.0"
asynctest = "^0.13.0"
pytest-dotenv = "^0.5.1"

[tool.poetry.scripts]
teached = "teached.manage:app"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["teached"]

[tool.coverage.report]
show_missing = true
fail_under = 84

[tool.isort]
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
line_length = 80

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
